+++ packages/groq/src/groq-chat-settings.ts
@@ -1,15 +1,19 @@
 // https://console.groq.com/docs/models
-// production models
 export type GroqChatModelId =
-  | 'deepseek-r1-distill-llama-70b'
+  // production models
   | 'gemma2-9b-it'
-  | 'gemma-7b-it'
   | 'llama-3.3-70b-versatile'
   | 'llama-3.1-8b-instant'
   | 'llama-guard-3-8b'
   | 'llama3-70b-8192'
   | 'llama3-8b-8192'
   | 'mixtral-8x7b-32768'
+  // preview models (selection)
+  | 'qwen-qwq-32b'
+  | 'mistral-saba-24b'
+  | 'qwen-2.5-32b'
+  | 'deepseek-r1-distill-qwen-32b'
+  | 'deepseek-r1-distill-llama-70b'
   | (string & {});
 
 export interface GroqChatSettings {