+++ packages/openai/src/convert-to-openai-chat-messages.ts
@@ -130,10 +130,6 @@ export function convertToOpenAIChatMessages({
               text += part.text;
               break;
             }
-            case 'redacted-reasoning':
-            case 'reasoning': {
-              break; // ignored
-            }
             case 'tool-call': {
               toolCalls.push({
                 id: part.toolCallId,
@@ -144,10 +140,6 @@ export function convertToOpenAIChatMessages({
                 },
               });
               break;
-            }
-            default: {
-              const _exhaustiveCheck: never = part;
-              throw new Error(`Unsupported part: ${_exhaustiveCheck}`);
             }
           }
         }