+++ packages/provider-utils/src/index.ts
@@ -12,8 +12,8 @@ export { loadOptionalSetting } from './load-optional-setting';
 export { loadSetting } from './load-setting';
 export * from './parse-json';
 export * from './post-to-api';
-export * from './resolve';
 export * from './remove-undefined-entries';
+export * from './resolve';
 export * from './response-handler';
 export * from './uint8-utils';
 export * from './validate-types';
@@ -21,5 +21,6 @@ export * from './validator';
 export * from './without-trailing-slash';
 
 export type { IDGenerator } from './generate-id';
+export { parseProviderOptions } from './parse-provider-options';
 export type { CoreToolCall, ToolCall } from './types/tool-call';
 export type { CoreToolResult, ToolResult } from './types/tool-result';