+++ packages/cohere/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ai-sdk/cohere
 
+## 1.1.18
+
+### Patch Changes
+
+- Updated dependencies [d0c4659]
+  - @ai-sdk/provider-utils@2.1.15
+
+## 1.1.17
+
+### Patch Changes
+
+- Updated dependencies [0bd5bc6]
+  - @ai-sdk/provider@1.0.12
+  - @ai-sdk/provider-utils@2.1.14
+
+## 1.1.16
+
+### Patch Changes
+
+- 4473677: feat (providers/cohere): add command-a model id to chat settings
+
 ## 1.1.15
 
 ### Patch Changes