+++ packages/deepinfra/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @ai-sdk/deepinfra
 
+## 0.1.18
+
+### Patch Changes
+
+- Updated dependencies [d0c4659]
+  - @ai-sdk/provider-utils@2.1.15
+  - @ai-sdk/openai-compatible@0.1.17
+
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies [0bd5bc6]
+  - @ai-sdk/provider@1.0.12
+  - @ai-sdk/openai-compatible@0.1.16
+  - @ai-sdk/provider-utils@2.1.14
+
 ## 0.1.16
 
 ### Patch Changes