+++ packages/provider-utils/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ai-sdk/provider-utils
 
+## 2.1.15
+
+### Patch Changes
+
+- d0c4659: feat (provider-utils): parseProviderOptions function
+
+## 2.1.14
+
+### Patch Changes
+
+- Updated dependencies [0bd5bc6]
+  - @ai-sdk/provider@1.0.12
+
 ## 2.1.13
 
 ### Patch Changes